Revert "feat(battlefield): add verdant_battlefield.glb and rename ResourceMap to Battlefield"
This reverts commit afa693446707fb7503550b90d08dfda4afec1c53.
