+++ apps/web/src/game/ResourceMap/Bush.tsx
