+++ apps/web/src/game/ResourceMap/Castle.tsx
