+++ apps/web/src/game/ResourceMap/PathPreview.tsx
