+++ apps/web/src/game/ResourceMap/PlayerPawn.tsx
