+++ apps/web/src/game/ResourceMap/TileHoverEffect.tsx
