+++ apps/web/src/pages/ResourceMapPage.css
